Include terminal functions from Term::ANSIScreen:
Term::ANSIScreen does not install properly because of Module::Package.
It will likely not see another release unless someone takes over
maintainership.

It is solely used here for a set of functions for moving on the screen
and getting terminal coloring. This commit introduces these functions
into the code, thus removing the dependency on Term::ANSIScreen.

Reference: RT #123497.